pyproject.toml
[project]
name = "diet-management"
version = "0.1.0"
description = "Telegram bot for diet tracking powered by LLMs"
requires-python = ">=3.11"
dependencies = [
    "python-telegram-bot[job-queue]>=21.0",
    "anthropic>=0.40.0",
    "openai>=1.50.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "sqlalchemy[asyncio]>=2.0",
    "aiosqlite>=0.20",
    "structlog>=24.0",
    "httpx>=0.27",
    "Pillow>=10.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "respx>=0.21",
    "mypy>=1.11",
    "ruff>=0.6",
    "pytest-socket>=0.7",
    "types-Pillow>=10.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["bot", "llm", "diet", "storage", "config"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "C4", "SIM"]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["S101"]

[tool.mypy]
strict = true
python_version = "3.11"
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = ["telegram.*", "structlog.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
